refactor: extract shared boundary-config reader
DEPENDENCY_CRUISER_CONFIG_NAMES + readBoundaryConfig were byte-identical in
architecture-fingerprint and architecture-monorepo — a real desync hazard
since both hash the boundary config into their fingerprints. Hoist to
utils/boundary-config.ts (standardizing on the readFileSafe form). Fingerprint
output unchanged (55 architecture tests green).
